Set LearningAgentSimulation to run in sync mode
This is crucial because AgentSimulationEvolution
should only access the data models within the
simulation after completion.
Rather than adding simulation completion
notifications to Presage.simulation for the thread
AgentSimulationEvolution to know about completion
this is a better approach.
Downside: can't really have UI if you run it on
the main thread.